moves Annotations.getDateRangeForPeriod to plugin class
